refactor: Extract `addRecipe` to `dbServices.ts`
- Add functionality to save recipe description to Indexed DB

References 186
