+++ components/organisms/RecipeForm.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
-import { nanoid } from 'nanoid';
-import { db } from '../../lib/db';
+import { addRecipe } from '../../services/dbService';
 
 import InputGroup from '../atoms/InputGroup';
 import SelectGroup from '../atoms/SelectGroup';
@@ -11,31 +10,19 @@ import IngredientInputs from '../molecules/IngredientInputs';
 const RecipeForm = () => {
   const [ status, setStatus ] = useState("");
 
-  const addRecipe = async (e) => {
+  const handleSubmit = async (e: { preventDefault: () => void; currentTarget: any; target: any; }) => {
     e.preventDefault();
-    
-    const form = e.currentTarget;
-    
-    try {
-      const formData = new FormData(form);
-      const formDataObject = Object.fromEntries(formData.entries());
 
-      const newRecipe = {
-        id: nanoid(),
-        name: formDataObject['recipe name'],
-        category: formDataObject['category'],
-      };
+    const { result, message } = await addRecipe(e);
 
-      await db.recipes.add(newRecipe);
-
-      setStatus(`Recipe "${newRecipe.name}" added!`);
-    } catch (err) {
-      console.error(err);
-    };
+    if (result !== 'error') e.target.reset();
+    setStatus(message);
   };
 
   return (
-    <form className="form" onSubmit={addRecipe}>
+    <form className="form" onSubmit={handleSubmit}>
+      <p>{status}</p>
+
       <fieldset className="form__fieldset">
         <legend className="bigger">Recipe Information</legend>
 
@@ -99,8 +86,6 @@ const RecipeForm = () => {
           modifier='--bad-job'
         />
       </div>
-
-      <p>{status}</p>
     </form>
   );
 };
